app/requirements/base.txt
fastapi
uvicorn

# for Form fastapi
python-multipart

# for jwt token
python-jose
passlib[bcrypt]

# for models
pydantic

# for async pymongo
motor

python-dotenv

redis

# for email signup
fastapi-mail
